requirements.txt
ipykernel
pandas
numpy
requests
websockets
python-dotenv
kafka-python
apache-airflow
redis
scikit-learn
mlflow
xgboost
psycopg2
questdb
boto3
minio
gunicorn
waitress
